Rename range-js to xpath-evaluate
